requirements.txt
Flask
flask-cors
google-generativeai
python-dotenv
PyMuPDF
python-docx